Added okHttp To Android Project
Created a new class called HttpRequest which does basic GET and POST requests all is need is the URL for the server which will usually be localhost:3000. I have also changed the manifest permisions to add ( android:usesCleartextTraffic="true" AND   <uses-permission android:name="android.permission.INTERNET"></uses-permission>) This allows for the App to connect to the server.